[rpm] Make rpm-libs an empty transitional package ; JB#46336
Move the shared libs from the -libs subpackage to the main package.
Leave the -libs subpackage empty except for a short README for now;
to be removed later once everything adjusts to having the libs in
the main package, and after some dependency cleanup.

The motivation for this is to avoid breakage that can occur when
something both rpm and rpm-libs depend on, e.g. liblua, gets updated,
and rpm and rpm-libs then updated on device, but not in a single
transaction.  This can lead to a situation where the installed rpm
and rpm-libs are linked against different versions of liblua, causing
rpm to segfault.

Signed-off-by: Aapi Hämäläinen <aapi.hamalainen@jollamobile.com>